% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_functions.R
\name{.pal_base}
\alias{.pal_base}
\title{Base function for color palette functions}
\usage{
.pal_base(namels, colors_df, with_names)
}
\arguments{
\item{namels}{list of quoted names (in `name_engl`) of colors to return}

\item{colors_df}{a data frame with cols `name_engl` and `hexcode`}

\item{with_names}{logical. Should color vector be returned with names?}
}
\value{
A character vector of hex colors, named depending on `with_names`
}
\description{
Wraps together the common parts for palette functions.
Different palettes are defined by the data frame argument.
}
\details{
`namels` is meant to be produced by `...` in actual palette functions.
}
