% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_functions.R
\name{pal_hsl}
\alias{pal_hsl}
\title{Get HSL hex color palette (or colors by name)}
\usage{
pal_hsl(..., with_names = TRUE)
}
\arguments{
\item{...}{One or multiple color names as character.
If empty, all colors are returned.}

\item{with_names}{Logical: return palette as named vector?}
}
\value{
A named vector of hex color codes.
}
\description{
Get HSL hex color palette (or colors by name)
}
\examples{
pal_hsl("train", "tram", "lightrail")
pal_hsl("train", "tram", "lightrail", with_names = FALSE)
pal_hsl()
}
