% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_functions.R
\name{pal_wsp}
\alias{pal_wsp}
\title{Get WSP hex color palette (or colors by name)}
\usage{
pal_wsp(..., with_names = TRUE)
}
\arguments{
\item{...}{One or multiple color names as character.
If empty, all colors are returned.}

\item{with_names}{Logical: return palette as named vector?}
}
\value{
A named vector of hex color codes.
}
\description{
Get WSP hex color palette (or colors by name)
}
\examples{
pal_wsp("red", "blue", "gray")
pal_wsp("red", "blue", "gray", with_names = FALSE)
pal_wsp()
}
