% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{show_palette}
\alias{show_palette}
\title{Plot palette colors}
\usage{
show_palette(pal)
}
\arguments{
\item{pal}{named character vector of hex color codes,
e.g. result of a `pal_*()` function}
}
\value{
a `ggplot` object
}
\description{
Plot palette colors
}
\examples{
show_palette(pal_hsl())

}
